smal refactorings to semantic model closure
